refactor: share native startup benchmark adapters
